# metrics: Reduce CPU cycles spent in Prometheus metric collection