Prior to this change, Prometheus description objects were re-created on
every collection cycle, even though label names change extremely rarely
in practice (perhaps even never on a given machine). This means that we
were unnecessarily causing CPU cycles. Using the benchmark in the
previous commit, we determine we're making collection at least 23%
faster, but based on production profiling data, we estimate this is
improving the Kubelet's baseline CPU usage by ~1.3% (potentially even
more since it also relieves the GC). This is because the benchmark shows
that all calls to `NewDesc` are completely disappearing from profiling
data, therefore stastically speaking reducing the kubelet's CPU time by
that much.

Production Kubelet CPU data: https://pprof.me/437533c/?filter_by_function=collectContainersInfo&search_string=collectContainersInfo
Before Optimizing: https://pprof.me/ac14fee/
After Optimizing: https://pprof.me/7ec5def/

```
benchstat old.txt new.txt
goos: linux
goarch: amd64
pkg: github.com/google/cadvisor/metrics
cpu: AMD Ryzen 5 3400GE with Radeon Vega Graphics
                      │   old.txt    │               new.txt               │
                      │    sec/op    │    sec/op     vs base               │
PrometheusCollector-8   614.4µ ± ∞ ¹   470.8µ ± ∞ ¹  -23.37% (p=0.008 n=5)
¹ need >= 6 samples for confidence interval at level 0.95
```

Side note: We believe we can further improve CPU usage here, but we think it requires a new API in the `prometheus.NewConstMetric` functionality, that allows caching of the result of `MakeLabelPairs`. We'll open a feature request for this.

This patch was created on the Polar Signals Let's Profile stream: https://www.youtube.com/watch?v=dkx-5vGCns8